refactor: reorganize imports and update paths for user and post types, enhancing code structure and maintainability
