+++ app/(without-sidebar)/(auth)/login/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { login } from "@/lib/utils/supabase/actions/auth/auth.actions";
-import { LoginSchemaErrorType, LoginSchemaType } from "@/lib/utils/supabase/validations/authSchema";
+import { LoginSchemaErrorType, LoginSchemaType } from "@/lib/validations/authSchema";
 import { useActionState } from "react";
 import FormWrapper from "../components/form-wrapper";
 import ErrorMessage from "@/components/error-message";
