+++ components/app-wrapper.tsx
@@ -1,4 +1,4 @@
-import UserProvider from "@/lib/context/user";
+import UserProvider from "@/lib/context/user.context";
 import { fetchUserProfile } from "@/lib/utils/supabase/actions/user/user.actions";
 import { createClient } from "@/lib/utils/supabase/server";
 import { ReactNode } from "react";
